+++ aspnet/RVTR.Account.DataContext/Repositories/IUnitOfWork.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-
-namespace RVTR.Account.DataContext.Repositories
-{
-  public interface IUnitOfWork
-  {
-    void Commit();
-  }
-}