+++ aspnet/RVTR.Account.DataContext/Repositories/UnitOfWork.cs
@@ -1,9 +1,27 @@
-using System;
+using System.Threading.Tasks;
+using RVTR.Account.ObjectModel.Models;
 
 namespace RVTR.Account.DataContext.Repositories
 {
-  public class UnitOfWork : IUnitOfWork
+  /// <summary>
+  /// Represents the _UnitOfWork_ repository
+  /// </summary>
+  public class UnitOfWork
   {
-    public void Commit() => throw new NotImplementedException();
+    private readonly AccountContext _context;
+
+    public Repository<AccountModel> Account { get; }
+    public Repository<ProfileModel> Profile { get; }
+
+    public UnitOfWork(AccountContext context)
+    {
+      _context = context;
+    }
+
+    /// <summary>
+    /// Represents the _UnitOfWork_ `Commit` method
+    /// </summary>
+    /// <returns></returns>
+    public async Task<int> CommitAsync() => await _context.SaveChangesAsync();
   }
 }