+++ aspnet/RVTR.Account.ObjectModel/Models/AccountModel.cs
@@ -12,7 +12,6 @@ namespace RVTR.Account.ObjectModel.Models
 
     public AddressModel Address { get; set; }
 
-    [Required]
     public string Name { get; set; }
 
     public IEnumerable<PaymentModel> Payments { get; set; }