+++ aspnet/RVTR.Account.UnitTesting/Tests/ProfileModelTest.cs
@@ -17,7 +17,9 @@ namespace RVTR.Account.UnitTesting.Tests
           Id = 0,
           Email = "email@email.com",
           Name = new NameModel(),
-          Phone = "1234567890"
+          Phone = "1234567890",
+          AccountId = null,
+          Account = null
         }
       }
     };