+++ aspnet/RVTR.Account.WebRpc/Services/AccountService.cs
@@ -11,9 +11,9 @@ namespace RVTR.Account.WebRpc
   public class AccountService : Account.AccountBase
   {
     private readonly ILogger<AccountService> _logger;
-    private readonly IUnitOfWork _unitOfWork;
+    private readonly UnitOfWork _unitOfWork;
 
-    public AccountService(ILogger<AccountService> logger, IUnitOfWork unitOfWork)
+    public AccountService(ILogger<AccountService> logger, UnitOfWork unitOfWork)
     {
       _logger = logger;
       _unitOfWork = unitOfWork;